Extract player capsule KCC step
