+++ src/ecs/systems/CollisionSystem.cpp
@@ -19,6 +19,7 @@
 #include "ecs/physics/TriMeshCollision.hpp"
 #include "ecs/systems/ExplosionSystem.hpp"
 #include "ecs/systems/FireSystem.hpp"
+#include "ecs/systems/KinematicCharacterController.hpp"
 
 #include <glm/geometric.hpp>
 
@@ -157,8 +158,7 @@ static void depenetrateBox(glm::vec3& pos, glm::vec3& vel, const glm::vec3& half
         const glm::vec3 boxCentre = (box.min + box.max) * 0.5f;
         const glm::vec3 contactPoint = pos - pushDir * std::abs(glm::dot(halfExtents, glm::abs(pushDir)));
         (void)boxCentre;
-        physics::debug::pushDepenContact(
-            contactPoint, pushDir, minPen, physics::debug::ContactSource::BoxDepen, 0);
+        physics::debug::pushDepenContact(contactPoint, pushDir, minPen, physics::debug::ContactSource::BoxDepen, 0);
     }
 }
 
@@ -358,101 +358,6 @@ depenetrate(glm::vec3& pos, glm::vec3& vel, const glm::vec3& halfExtents, const 
     }
 }
 
-// Modern two-capsule character controller (Havok / Jolt / Unity-KCC pattern)
-//
-// Replaces:
-//   * the legacy AABB lift→horiz→drop stair-step dance (`tryStepUp`),
-//     which broke once the player became a true capsule because the
-//     drop sweep's bounded-reach check rejects edge-corner contacts
-//     that an AABB happened to reach,
-//   * the per-feature MTV-summing capsule depen, whose front+back
-//     contacts on thin geometry cancelled to zero push, and
-//   * the CA-loop "contact branch", which pushed the player by a fixed
-//     `k_pushback` per iteration in whatever direction the BVH happened
-//     to return — causing the slow downward drift through thin floors.
-//
-// The new model:
-//
-//   * Per-pass-deepest-first depen (`depenetrateCapsuleVsWorld`) picks
-//     the single deepest Voronoi contact across the whole world and
-//     ejects fully; oscillation between opposing normals trips the
-//     `emergencyUnstick` fallback (free-space probe in cardinal dirs).
-//   * Horizontal motion uses the *walk capsule* (foot above stepHeight),
-//     so stair risers / curbs / thresholds within stepHeight are
-//     physically invisible to the sweep — they cannot be hit.
-//   * Ground resolve (`resolveGround`) does ONE downward swept query for
-//     the *full capsule* and snaps the foot onto whatever walkable
-//     surface is below.  This is what climbs stairs (foot snaps to
-//     tread), descends slopes (foot snaps to lower slope), and detects
-//     "grounded" — all from a single query.
-//   * Airborne vertical motion stays a swept full-capsule query, so
-//     mid-air clipping into low walls / ceilings is honest.
-
-/// @brief Build a `physics::CapsuleShape` query from a player's
-/// `CollisionShape`.  Capsule axis is `+Y` when gravity is normal and
-/// `-Y` when gravity is flipped — `probeGround` shoots its sweep along
-/// `-capsule.up`, so flipping `up` flips the floor-search direction
-/// without any other code change.  The capsule's Minkowski math is
-/// symmetric under reflection of `up`, so swept queries on the flipped
-/// capsule are numerically identical to the un-flipped queries.
-static physics::CapsuleShape makeCapsuleQuery(const CollisionShape& shape, bool gravityFlipped)
-{
-    return physics::CapsuleShape{
-        .radius = shape.radius,
-        .halfHeight = shape.halfHeight,
-        .up = glm::vec3{0.0f, gravityFlipped ? -1.0f : 1.0f, 0.0f},
-    };
-}
-
-/// @brief Settle the player onto the closest walkable surface within
-/// `maxSnapDistance` along the foot direction (`-capsule.up`).
-///
-/// Unified replacement for the legacy `Phase 2 slope-stick` +
-/// `Phase 3 ground probe` + `tryStepUp drop sweep` — one swept query,
-/// one snap, one source of truth.  Stair climbs naturally: the
-/// preceding horizontal walk-capsule sweep slid past the riser; the
-/// ground probe here uses the *full* capsule and finds the tread
-/// below; the snap places the foot on the tread.
-///
-/// On success: sets `state.grounded = true`, captures the ground
-/// normal, zeroes only the gravity-axis component of velocity
-/// (preserves tangential slide), and adjusts `pos` along
-/// `capsule.up` so the foot rests `k_pushback` above the surface.
-///
-/// Returns true when the player was settled on walkable ground.
-static bool resolveGround(physics::CapsuleShape capsule,
-                          glm::vec3& pos,
-                          glm::vec3& vel,
-                          PlayerVisState& state,
-                          float maxSnapDistance,
-                          const physics::WorldGeometry& world)
-{
-    const physics::GroundProbeResult probe = physics::probeGround(capsule, pos, maxSnapDistance, world);
-    if (!probe.hit || !probe.walkable)
-        return false;
-
-    // Settle the capsule along the gravity axis so its surface touches the
-    // ground point with `k_pushback` separation along the contact normal.
-    // On a slope the contact normal is not parallel to `capsule.up`, so the
-    // along-axis offset is `dot(normal, up) * (pushback + minkExt(normal))`
-    // — using `halfHeight + radius` directly would float the player off
-    // angled slopes by `(1 - dot(normal,up)) * minkExt` units.
-    const float k_minkExtAlongNormal = capsule.minkowskiExtent(probe.normal);
-    const float k_targetAlongUp =
-        glm::dot(probe.point, capsule.up) + glm::dot(probe.normal, capsule.up) * (k_pushback + k_minkExtAlongNormal);
-    const float k_currentAlongUp = glm::dot(pos, capsule.up);
-    pos += capsule.up * (k_targetAlongUp - k_currentAlongUp);
-
-    state.grounded = true;
-    state.groundNormal = probe.normal;
-    if (!state.grappleActive) {
-        const float k_vAlongUp = glm::dot(vel, capsule.up);
-        if (k_vAlongUp < 0.0f)
-            vel -= capsule.up * k_vAlongUp;
-    }
-    return true;
-}
-
 void runCollision(Registry& registry, float dt, const physics::WorldGeometry& world)
 {
     // CollisionSystem only reads/writes the replicated half (grounded /
@@ -479,243 +384,9 @@ void runCollision(Registry& registry, float dt, const physics::WorldGeometry& wo
         auto& vel = registry.get<Velocity>(e);
         const auto& shape = registry.get<CollisionShape>(e);
         auto& state = registry.get<PlayerVisState>(e);
-        {
-            // --- per-tick state setup ---
-            const bool k_wasGrounded = state.grounded;
-            state.grounded = false;
-
-            const physics::CapsuleShape capsule = makeCapsuleQuery(shape, state.gravityFlipped);
-            const glm::vec3 worldUp = capsule.up; // direction OPPOSITE gravity
-
-            const bool diagOn = physics::diag::isEnabled();
-            physics::diag::PlayerFrame diagFrame{};
-            if (diagOn) {
-                diagFrame.entity = e;
-                diagFrame.posBefore = pos.value;
-                diagFrame.velBefore = vel.value;
-            }
-
-            // --- Phase 0: depenetration (per-pass-deepest single-MTV resolver) ---
-            //
-            // Replaces the legacy per-feature MTV-summing depen.  Symmetric
-            // contacts on thin geometry no longer cancel to zero push: we
-            // pick the deepest single violator, eject fully, and re-probe.
-            // Oscillation between opposing normals (genuinely ambiguous
-            // two-sided geometry) trips the emergency-unstick fallback.
-            physics::depenetrateCapsuleVsWorld(pos.value, vel.value, capsule, world);
-
-            if (diagOn) {
-                diagFrame.posAfterDepen = pos.value;
-                diagFrame.depenPushDistance = glm::length(pos.value - diagFrame.posBefore);
-                if (diagFrame.depenPushDistance > 20.0f)
-                    diagFrame.flags |= physics::diag::PhaseFlag::DeepPenetration;
-            }
-
-            // --- Phase 1: motion integration ---
-            //
-            // Grounded ambulation: walk-capsule sweep over horizontal
-            // velocity only.  The walk capsule's foot sits stepHeight above
-            // the player's foot, so steps / risers within stepHeight are
-            // physically invisible to the sweep — they cannot be hit.
-            //
-            // Airborne or grappling: full-capsule sweep over full 3D
-            // velocity.  Feet must clear low walls in mid-air, and the
-            // grapple cable should pull the player honestly through 3D
-            // space.
-            const bool useWalkCapsule = k_wasGrounded && !state.grappleActive;
-
-            const physics::CapsuleShape sweepCapsule =
-                useWalkCapsule ? capsule.walkShape(physics::k_stepHeight) : capsule;
-            const glm::vec3 sweepCenterOffset =
-                useWalkCapsule ? capsule.walkCenterOffset(physics::k_stepHeight) : glm::vec3{0.0f};
-
-            glm::vec3 phaseVel = vel.value;
-            if (useWalkCapsule) {
-                // Strip the gravity-axis component; ground resolve below
-                // settles the foot onto whatever's there.
-                phaseVel -= worldUp * glm::dot(phaseVel, worldUp);
-            }
-
-            // Sub-step bound: cap per-substep motion at radius * safetyRatio
-            // so high-velocity grapple yanks split into multiple substeps.
-            const float fullStep = glm::length(phaseVel) * dt;
-            const float maxSafeStep = sweepCapsule.radius * physics::k_substepSafetyRatio;
-            const int numSubsteps =
-                (physics::k_enableSubstepping && maxSafeStep > 1e-6f && fullStep > maxSafeStep)
-                    ? std::min(physics::k_maxSubsteps,
-                               static_cast<int>(std::ceil(fullStep / maxSafeStep)))
-                    : 1;
-            const float subDt = dt / static_cast<float>(numSubsteps);
-            constexpr int k_maxCAIterations = 8;
-            bool caExhaustedAnySubstep = false;
-
-            for (int sub = 0; sub < numSubsteps; ++sub) {
-                float remainingTime = subDt;
-                int iter = 0;
-                for (; iter < k_maxCAIterations && remainingTime > 1e-5f; ++iter) {
-                    const glm::vec3 sweepStart = pos.value + sweepCenterOffset;
-                    const float vLen = glm::length(phaseVel);
-                    const float motionBound = vLen * remainingTime;
-                    if (motionBound < 1e-6f) {
-                        remainingTime = 0.0f;
-                        break;
-                    }
-
-                    // Open-space fast-reject — omni-clearance lets us skip
-                    // the sweep entirely when no geometry is within reach.
-                    const physics::ClearanceResult clr =
-                        physics::clearanceCapsuleVsWorld(sweepCapsule, sweepStart, world);
-                    if (clr.distance > motionBound + k_pushback) {
-                        pos.value += phaseVel * remainingTime;
-                        remainingTime = 0.0f;
-                        break;
-                    }
-
-                    // Sweep TOI for exact directional contact.
-                    const glm::vec3 sweepEnd = sweepStart + phaseVel * remainingTime;
-                    const physics::HitResult hit =
-                        physics::sweepAll(sweepCapsule, sweepStart, sweepEnd, world);
-                    if (!hit.hit) {
-                        pos.value += phaseVel * remainingTime;
-                        remainingTime = 0.0f;
-                        break;
-                    }
-
-                    pos.value += phaseVel * hit.tFirst * remainingTime;
-                    remainingTime *= (1.0f - hit.tFirst);
-
-                    if (diagOn) {
-                        ++diagFrame.bumpHits;
-                        diagFrame.lastHitNormal = hit.normal;
-                    }
-                    if (physics::debug::isEnabled()) {
-                        const float r = sweepCapsule.minkowskiExtent(hit.normal);
-                        physics::debug::pushSweepContact(pos.value + sweepCenterOffset - hit.normal * r,
-                                                         hit.normal,
-                                                         physics::debug::ContactSource::PlaneSweep);
-                    }
-
-                    // Clip + pushback at the hit.  In the grounded /
-                    // walk-capsule branch every hit is a wall (the walk
-                    // capsule's foot cannot reach a floor), so we use the
-                    // wall overbounce uniformly.  In the airborne / full-
-                    // capsule branch we still treat floor hits as floors
-                    // (settle grounded) for the same-tick landing case.
-                    if (!useWalkCapsule) {
-                        const bool landed = glm::dot(hit.normal, worldUp) >= physics::k_floorAngleCos;
-                        if (landed) {
-                            state.grounded = true;
-                            state.groundNormal = hit.normal;
-                            phaseVel = physics::clipVelocity(phaseVel, hit.normal, physics::k_overbounceFloor);
-                        } else {
-                            phaseVel = physics::clipVelocity(phaseVel, hit.normal, physics::k_overbounceWall);
-                        }
-                    } else {
-                        phaseVel = physics::clipVelocity(phaseVel, hit.normal, physics::k_overbounceWall);
-                    }
-                    pos.value += hit.normal * k_pushback;
-                }
-                if (iter >= k_maxCAIterations && remainingTime > 1e-5f)
-                    caExhaustedAnySubstep = true;
-            }
-
-            // Write back the (possibly clipped) velocity.
-            if (useWalkCapsule) {
-                vel.value.x = phaseVel.x;
-                vel.value.z = phaseVel.z;
-                // vel.y preserved — gravity component is handled by Phase 2.
-            } else {
-                vel.value = phaseVel;
-            }
-
-            // --- Phase 2: ground resolve ---
-            //
-            // One downward swept query with the full capsule.  When the
-            // player was grounded last tick, the probe extends to
-            // (effectiveStep + groundSnap) so it catches both step-ups
-            // (after a walk-capsule sweep over a riser) and step-downs
-            // (descending stairs / slopes).  When airborne, the probe is
-            // just a tiny landing check.
-            //
-            // Skipped during grapple: the player flies freely along the
-            // cable without being snapped to the ground.
-            //
-            // Skipped while actively jumping (vy strongly opposing gravity):
-            // we don't want to swallow a fresh jump impulse.
-            if (!state.grappleActive) {
-                const float k_vAlongUp = glm::dot(vel.value, worldUp);
-                const bool jumping = k_vAlongUp > 10.0f;
-                if (!jumping) {
-                    const float effStep = capsule.effectiveStepHeight(physics::k_stepHeight);
-                    const float snap = k_wasGrounded
-                                           ? (effStep + physics::k_groundSnapDistance)
-                                           : physics::k_groundSnapDistance;
-                    resolveGround(capsule, pos.value, vel.value, state, snap, world);
-                }
-            }
-
-            // --- Phase 3: vertical motion when still airborne ---
-            //
-            // Grounded case is settled by Phase 2.  Airborne case (free-
-            // fall, mid-jump, mid-grapple-arc) integrates the gravity-axis
-            // velocity with a full-capsule swept query so feet honestly
-            // clip into low walls / ceilings during vertical motion.
-            //
-            // (Horizontal was already integrated in Phase 1; this is just
-            // the Y component.)
-            if (!state.grounded && !state.grappleActive) {
-                const float k_vAlongUp = glm::dot(vel.value, worldUp);
-                const glm::vec3 vMotion = worldUp * k_vAlongUp;
-                if (glm::dot(vMotion, vMotion) > 1e-12f) {
-                    const glm::vec3 vTarget = pos.value + vMotion * dt;
-                    const physics::HitResult vHit = physics::sweepAll(capsule, pos.value, vTarget, world);
-                    if (!vHit.hit) {
-                        pos.value = vTarget;
-                    } else {
-                        pos.value += vMotion * vHit.tFirst * dt;
-                        pos.value += vHit.normal * k_pushback;
-                        const bool landed = glm::dot(vHit.normal, worldUp) >= physics::k_floorAngleCos;
-                        if (landed) {
-                            state.grounded = true;
-                            state.groundNormal = vHit.normal;
-                            vel.value = physics::clipVelocity(vel.value, vHit.normal, physics::k_overbounceFloor);
-                        } else {
-                            vel.value = physics::clipVelocity(vel.value, vHit.normal, physics::k_overbounceWall);
-                        }
-                    }
-                }
-            }
-
-            // --- Diag finalize ---
-            if (diagOn) {
-                diagFrame.posAfter = pos.value;
-                diagFrame.velAfter = vel.value;
-                if (caExhaustedAnySubstep)
-                    diagFrame.flags |= physics::diag::PhaseFlag::BumpExhausted;
-                if (state.grounded)
-                    diagFrame.flags |= physics::diag::PhaseFlag::Grounded;
-                if (state.grappleActive)
-                    diagFrame.flags |= physics::diag::PhaseFlag::GrappleActive;
-                if (state.gravityFlipped)
-                    diagFrame.flags |= physics::diag::PhaseFlag::GravityFlipped;
-                diagFrame.moveMode = static_cast<int>(state.moveMode);
-                diagFrame.wallrunSide = static_cast<int>(state.wallRunSide);
-                diagFrame.jumpCount = state.jumpCount;
-                if (diagFrame.moveMode == 2) // WallRunning
-                    diagFrame.flags |= physics::diag::PhaseFlag::WallRunning;
-                if (diagFrame.moveMode == 1) // Sliding
-                    diagFrame.flags |= physics::diag::PhaseFlag::Sliding;
-                if (diagFrame.moveMode == 3) // Climbing
-                    diagFrame.flags |= physics::diag::PhaseFlag::Climbing;
-                if (diagFrame.moveMode == 4) // LedgeGrabbing
-                    diagFrame.flags |= physics::diag::PhaseFlag::LedgeGrabbing;
-                if (auto* sim = registry.try_get<PlayerSimState>(e); sim != nullptr) {
-                    if (sim->jumpedThisTick)
-                        diagFrame.flags |= physics::diag::PhaseFlag::DoubleJumped;
-                }
-                physics::diag::recordFrame(diagFrame);
-            }
-        }
+        const PlayerSimState* sim = registry.try_get<PlayerSimState>(e);
+        const bool jumpedThisTick = sim != nullptr && sim->jumpedThisTick;
+        runKinematicCharacterController(pos.value, vel.value, shape, state, dt, world, e, jumpedThisTick);
     };
 
 #if GROUP2_COLLISION_HAS_PARALLEL
